docs: moves conversion example from Getting Started to More Examples
